+++ Clicked/Config2/Pages/Binding.lua
@@ -17,6 +17,7 @@
 --- @class BindingConfigTab
 --- @field public container AceGUIContainer
 --- @field public bindings Binding[]
+--- @field public controller BindingConfigBindingPage
 --- @field public Show? fun(self: BindingConfigTab, container: AceGUIContainer)
 --- @field public Hide? fun(self: BindingConfigTab)
 --- @field public Redraw? fun(self: BindingConfigTab, container: AceGUIContainer, binding: Binding)
@@ -134,6 +135,7 @@ Addon.BindingConfig.BindingPage = {
 	filteredTargets = {}
 }
 
+--- @protected
 function Addon.BindingConfig.BindingPage:Hide()
 	local currentTab = self.currentTab
 
@@ -143,6 +145,7 @@ function Addon.BindingConfig.BindingPage:Hide()
 
 		tab.container = nil
 		tab.bindings = nil
+		tab.controller = nil
 
 		self.currentTab = nil
 	end
@@ -152,6 +155,7 @@ function Addon.BindingConfig.BindingPage:Hide()
 	self.tabWidget = nil
 end
 
+--- @protected
 function Addon.BindingConfig.BindingPage:Redraw()
 	do
 		--- @param binding Binding
@@ -219,6 +223,7 @@ function Addon.BindingConfig.BindingPage:Redraw()
 	self:CreateTabGroup()
 end
 
+--- @protected
 function Addon.BindingConfig.BindingPage:OnBindingReload()
 	self:UpdateTabGroup()
 
@@ -227,6 +232,18 @@ function Addon.BindingConfig.BindingPage:OnBindingReload()
 	if currentTab ~= nil then
 		local impl = self.tabs[currentTab].implementation
 		Addon:SafeCall(impl.OnBindingReload, impl)
+	end
+end
+
+--- Redraw the currently active tab
+function Addon.BindingConfig.BindingPage:RedrawTab()
+	local currentTab = self.currentTab
+
+	if currentTab ~= nil then
+		local impl = self.tabs[currentTab].implementation
+
+		self.tabWidget:ReleaseChildren()
+		Addon:SafeCall(impl.Redraw, impl)
 	end
 end
 
@@ -245,6 +262,7 @@ function Addon.BindingConfig.BindingPage:ActivateTabGroup(group)
 
 		tab.container = nil
 		tab.bindings = nil
+		tab.controller = nil
 
 		self.currentTab = nil
 	end
@@ -257,22 +275,9 @@ function Addon.BindingConfig.BindingPage:ActivateTabGroup(group)
 
 	impl.container = self.tabWidget
 	impl.bindings = self.filteredTargets[group] or self.targets
+	impl.controller = self
 
 	self:RedrawTab()
-end
-
---- Redraw the currently active tab
----
---- @private
-function Addon.BindingConfig.BindingPage:RedrawTab()
-	local currentTab = self.currentTab
-
-	if currentTab ~= nil then
-		local impl = self.tabs[currentTab].implementation
-
-		self.tabWidget:ReleaseChildren()
-		Addon:SafeCall(impl.Redraw, impl)
-	end
 end
 
 --- Update the available tabs in the tab group widget